40 mb of data guaranteed
Instructions:
- Run StmApiFetch.py
- Run VilleDeMontrealFetch.py

+++ HelperFunctions.py
@@ -5,6 +5,7 @@ import csv
 import requests
 import json
 import gtfs_pb2
+import os
 
 day_of_week_mapping = {
     "Lundi": "Monday",
@@ -292,7 +293,7 @@ def extract_line_name(line_string):
     
 def epoch_to_date(epoch):
     epoch = int(epoch)
-    return datetime.fromtimestamp(epoch).strftime('%d-%m-%Y')
+    return datetime.fromtimestamp(epoch).strftime('%m-%d-%Y')
 
 def epoch_to_timestamp(epoch):
     try:
@@ -407,12 +408,40 @@ def fetch_and_create_json_stm_response_json(url):
                 
                 # Convert the Protobuf message to JSON
                 json_trip_updates = MessageToJson(trip_updates)
+
+                # Generate the filename based on the current month and date
+                current_date = datetime.now()
+                filename = f"stm_response_trips_{current_date.month}_{current_date.day}.json"
                 
-                # Write the JSON data to stm_response_trips.json
-                with open('stm_response_trips.json', 'w') as json_file:
-                    json.dump(json.loads(json_trip_updates), json_file, indent=4)
+                # Ensure the LiveInformation directory exists
+                directory = "LiveInformation"
+                if not os.path.exists(directory):
+                    os.makedirs(directory)
                 
-                print("Protobuf response has been written to stm_response_trips.json")
+                # Full path to the file
+                filepath = os.path.join(directory, filename)
+                
+                # Check if the file already exists and is listed in live_trip_dates.txt
+                file_exists = os.path.exists(filepath)
+                file_listed = False
+                if os.path.exists('live_trip_dates.txt'):
+                    with open('live_trip_dates.txt', 'r') as date_file:
+                        file_listed = filepath in date_file.read().splitlines()
+                
+                if not file_exists or not file_listed:
+                    # Write the JSON data to the generated filename in the LiveInformation directory
+                    with open(filepath, 'w') as json_file:
+                        json.dump(json.loads(json_trip_updates), json_file, indent=4)
+                    
+                    print(f"Protobuf response has been written to {filepath}")
+                    
+                    # Write the filename into the live_trip_dates.txt file
+                    with open('live_trip_dates.txt', 'a') as date_file:
+                        date_file.write(f"{filepath}\n")
+                        date_file.close()
+                else:
+                    print(f"File {filepath} already exists and is listed in live_trip_dates.txt. Skipping write.")
+                
             except Exception as e:
                 print(f"Error: {e}")
     else:
